tests/requirements.txt
locust>=2.20.0
requests>=2.31.0
faker>=22.0.0
pyyaml>=6.0.1
